+++ glamod_marine_processing/__init__.py
@@ -4,3 +4,4 @@ from __future__ import annotations
 
 from . import obs_suite, qc_suite  # noqa
 from .pre_processing import pre_processing  # noqa
+from .post_processing import post_processing  # noqa