+++ glamod_marine_processing/post_processing/__init__.py
@@ -1,0 +1,5 @@
+"""GLAMOD marine processing post-processing package."""
+
+from __future__ import annotations
+
+from .post_processing import post_processing  # noqa